feat: expand metadata (#63)
- This PR changes metadata "leaf types" from dict[str, list[str]] to a
new MetadataType, which at the moment contains value (the field
extraction value), experimental_confidence (an experimental version of
confidence score for the field extraction result), and chunk_references
(chunk ids from which the field extraction result was gathered)
- Also included a helpful utils function for getting specific chunks
from chunk ids.
- Updated GitHub workflow files so that pushes to main don't release new
versions of the library, only pushes to the "release" branch

Some open questions:
- Should 'experimental_confidence' be called something else? Or, should
we even consider the more complex approach of having a sort of
"beta_parse" function or even a beta module (beta.parse)? Note that this
would likely entail traversing the metadata and removing confidence from
the original parse function.
- Should we add more utils functions? Or should this function be changed
to take in a list of chunk ids and return a list of chunks? Or something
else?

Note: As of the time I am making this PR, confidence and value aren't
added in the API, so I'm keeping them as optional in the MetadataType
definition. Eventually, they shouldn't be.