chore: dont break publishing pipeline if version bump fails to commit (#75)
Although pypi publication succeeds, the publishing pipeline is marked as
failed if the version in `pyproject.toml` in main is already set to the
published version.

This PR fixes that by ignoring commit / push errors. The version in
`pyproject.toml` is not that important too, as we generate the version
number when building the GitHub release.